requirements.txt
numpy
pandas
flask
requests
pdfquery
tabula-py
json2html
fuzzywuzzy
tabula-py